+++ app/controllers/analyses_controller.rb
@@ -1,23 +1,37 @@
 class AnalysesController < ApplicationController
+  before_action :check_request_analysis, only: %i[show]
+
   def manage
     @analysis = Analysis.new
     @analyses = Analysis.all.order(created_at: :desc).page(params[:page])
   end
 
   def execute
-    check_absent_param(execute_params, %i[num_data num_tree])
+    check_schema(execute_schema, execute_params, 'analysis')
 
     analysis = Analysis.new(execute_params)
+    analysis.build_result
     unless analysis.save
-      error_codes = analysis.errors.messages.keys.map {|key| "invalid_param_#{key}" }
-      raise BadRequest, error_codes
+      raise BadRequest, messages: analysis.errors.messages, resource: 'analysis'
     end
 
     AnalysisJob.perform_later(analysis.id)
     render status: :ok, json: {}
   end
 
+  def show
+    @analysis = request_analysis
+  end
+
   private
+
+  def check_request_analysis
+    raise NotFound unless request_analysis
+  end
+
+  def request_analysis
+    @request_analysis ||= Analysis.find_by(request.path_parameters.slice(:analysis_id))
+  end
 
   def execute_params
     @execute_params ||= request.request_parameters.slice(
@@ -25,5 +39,17 @@ class AnalysesController < ApplicationController
       :num_tree,
       :num_entry,
     )
+  end
+
+  def execute_schema
+    @execute_schema ||= {
+      type: :object,
+      required: %i[num_data num_tree],
+      properties: {
+        num_data: {type: :string, pattern: '^[1-9][0-9]*$'},
+        num_tree: {type: :string, pattern: '^[1-9][0-9]*$'},
+        num_entry: {type: :string, pattern: '^([1-9][0-9]*|\s*)$'},
+      },
+    }
   end
 end