+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 4.14.3 (2020/09/02)
+- [UPDATE] gems for memory leak
+- [UPDATE] gems
+
+# 4.14.2 (2020/08/29)
+- [UPDATE] config/settings.yml to add racewise features
+
+# 4.14.1 (2020/08/22)
+- [UPDATE] gems
+- [UPDATE] error spec by json-schema
+
+# 4.14.0 (2020/07/24)
+- [UPDATE] analyses view to confirm result
+- [UPDATE] gems
+
 # 4.13.0 (2020/04/29)
 - [UPDATE] views to reload job info automatically
 - [UPDATE] gems